requirements.txt
Django==4.2.4
djangorestframework==3.14.0
psycopg2-binary==2.9.10
drf-yasg==1.21.4
gunicorn==21.2.0
django-heroku==0.3.1
whitenoise==6.5.0